requirements.txt
annotated-types==0.7.0
ansicon==1.89.0
anyio==4.9.0
blessed==1.21.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
editor==1.6.6
google-auth==2.40.2
google-genai==1.18.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
inquirer==3.4.0
jinxed==1.3.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.5
pydantic_core==2.33.2
readchar==4.2.1
requests==2.32.3
rsa==4.9.1
runs==1.2.2
sniffio==1.3.1
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
wcwidth==0.2.13
websockets==15.0.1
xmod==1.8.1

google-generativeai
streamlit
PyPDF2
python-dotenv
inquirer
